+++ src/types/IContent.ts
@@ -6,13 +6,13 @@
 import { Readable } from 'node:stream'
 //
 import { DataTable } from "./DataTable"
-import { TFilesDataProviderOptions } from "../providers/data/FilesDataProvider"
+import { TContentConfig } from "../providers/data/FilesDataProvider"
 import { VirtualFileSystem } from "../utils/VirtualFileSystem "
 
 
 export interface IContent {
     EntityName: string
-    Options: TFilesDataProviderOptions
+    Options: TContentConfig
     Content: VirtualFileSystem
     Init(name: string, content: Readable): void
     Get(sqlQuery?: string): Promise<DataTable>